* ColorDictionary class creation and main.js refactor
